Lockscreen : should fix tests
